requirements.txt
torch
openai-whisper
sounddevice
numpy
requests
scipy
